Prevent wireless cameras from being deleted from default_project

+++ src/task/camera.ts
@@ -13,45 +13,40 @@ export async function UpdateSerialNumber(task: TaskInput<gql.UpdateCameraSerialN
 export async function DeleteCamera(task: TaskInput<gql.DeleteCameraInput>) {
   const context = { user: { is_superuser: true, curr_project: task.projectId } as User };
   console.log('CameraModel.deleteCameraConfig - input: ', task.config);
+  const { cameraId } = task.config;
   const errors = [];
-  try {
-    // Step 1: delete deployments from views
-    await ProjectModel.removeCameraFromViews(
-      {
-        cameraId: task.config.cameraId,
-      },
-      context,
-    );
-    // Step 2: delete camera record from project
-    await ProjectModel.deleteCameraConfig(
-      {
-        cameraId: task.config.cameraId,
-      },
-      context,
-    );
 
-    // Step3: delete images associated with this camera
-    const deleteRes = await DeleteImagesByFilter({
-      projectId: task.projectId,
-      config: {
-        filters: {
-          cameras: [task.config.cameraId],
-        },
-      },
-      type: 'DeleteImagesByFilter',
-      user: task.user,
-    });
-    if (deleteRes.errors) {
-      errors.push(...deleteRes.errors);
-    }
-    // Step 4: unregister camera
-    if (
-      (await CameraModel.getWirelessCameras({ _ids: [task.config.cameraId] }, context)).length > 0
-    ) {
-      await CameraModel.removeProjectRegistration({ cameraId: task.config.cameraId }, context);
-    }
-  } catch (err) {
-    return { isOk: false, error: err };
+  // Prevent deleting wireless cameras from default_project
+  const wirelessCam = await CameraModel.getWirelessCameras({ _ids: [cameraId] }, context);
+  const isWirelessCam = wirelessCam.length > 0;
+  if (isWirelessCam && task.projectId === 'default_project') {
+    throw new Error('You cannot delete wireless cameras from the Default Project');
   }
+
+  // Step 1: delete deployments from views
+  await ProjectModel.removeCameraFromViews({ cameraId }, context);
+
+  // Step 2: delete camera record from project
+  await ProjectModel.deleteCameraConfig({ cameraId }, context);
+
+  // Step 3: delete images associated with this camera
+  const deleteRes = await DeleteImagesByFilter({
+    projectId: task.projectId,
+    config: { filters: { cameras: [cameraId] } },
+    type: 'DeleteImagesByFilter',
+    user: task.user,
+  });
+
+  if (deleteRes.errors) {
+    errors.push(...deleteRes.errors);
+  }
+
+  // Step 4: if wireless camera, unregister camera
+  if (isWirelessCam) {
+    await CameraModel.removeProjectRegistration({ cameraId }, context);
+  }
+
+  // Note: `errors` is for any errors that occurred during the delete images step
+  // the task will be marked as complete and the errors are not displayed to the user
   return { isOk: true, errors: errors };
 }
